+++ src/com/sun/ts/tests/ejb30/lite/stateful/concurrency/accesstimeout/jsfdescriptor/build.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <!--
 
-    Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved.
+    Copyright (c) 2022 Oracle and/or its affiliates. All rights reserved.
 
     This program and the accompanying materials are made available under the
     terms of the Eclipse Public License v. 2.0, which is available at
@@ -29,9 +29,10 @@ com/sun/ts/tests/ejb30/common/helper/ServiceLocator.class,
 com/sun/ts/tests/ejb30/lite/stateful/concurrency/common/Pinger.class,
 com/sun/ts/tests/ejb30/lite/stateful/concurrency/common/StatefulConcurrencyIF.class,
 com/sun/ts/tests/ejb30/lite/stateful/concurrency/common/StatefulConcurrencyClientBase.class,
+com/sun/ts/tests/ejb30/lite/stateful/concurrency/common/StatefulConcurrencyJsfClientBase.class,
 
 com/sun/ts/tests/ejb30/lite/stateful/concurrency/accesstimeout/common/AccessTimeoutIF.class,
-com/sun/ts/tests/ejb30/lite/stateful/concurrency/accesstimeout/common/ClientBase.class,
+com/sun/ts/tests/ejb30/lite/stateful/concurrency/accesstimeout/common/JsfClientBase.class,
 com/sun/ts/tests/ejb30/lite/stateful/concurrency/accesstimeout/common/PlainAccessTimeoutBeanBase.class
         ">
         </ts.vehicles>
