+++ src/com/sun/ts/tests/ejb30/lite/stateful/concurrency/accesstimeout/jsfdescriptor/ejb-jar.xml
@@ -1,0 +1,60 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    Copyright (c) 2022 Oracle and/or its affiliates. All rights reserved.
+
+    This program and the accompanying materials are made available under the
+    terms of the Eclipse Public License v. 2.0, which is available at
+    http://www.eclipse.org/legal/epl-2.0.
+
+    This Source Code may also be made available under the following Secondary
+    Licenses when the conditions for such availability set forth in the
+    Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
+    version 2 with the GNU Classpath Exception, which is available at
+    https://www.gnu.org/software/classpath/license.html.
+
+    SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+
+-->
+
+<ejb-jar xmlns="https://jakarta.ee/xml/ns/jakartaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    version="4.0" xsi:schemaLocation="https://jakarta.ee/xml/ns/jakartaee https://jakarta.ee/xml/ns/jakartaee/ejb-jar_4_0.xsd">
+    <enterprise-beans>
+        <session>
+            <ejb-name>BeanClassLevelAccessTimeoutBean</ejb-name>
+            <concurrent-method>
+                <method>
+                    <method-name>beanClassLevel</method-name>
+                </method>
+                <lock>Write</lock>
+                <access-timeout>
+                    <!-- timeout value should equal to BEAN_CLASS_LEVEL_TIMEOUT_MILLIS in AccessTimeoutIF -->
+                    <timeout>5000</timeout>
+                    <unit>Milliseconds</unit>
+                </access-timeout>
+            </concurrent-method>
+            <concurrent-method>
+                <method>
+                    <method-name>beanClassLevel2</method-name>
+                </method>
+                <lock>Write</lock>
+                <access-timeout>
+                    <timeout>5000000</timeout>
+                    <unit>Microseconds</unit>
+                </access-timeout>
+            </concurrent-method>
+            <concurrent-method>
+                <method>
+                    <method-name>ping</method-name>
+                    <method-params/>
+                </method>
+                <lock>Write</lock>
+                <access-timeout>
+                    <timeout>5000000000</timeout>
+                    <unit>Nanoseconds</unit>
+                </access-timeout>
+            </concurrent-method>
+        </session>
+
+    </enterprise-beans>
+</ejb-jar>
